benchmarks/requirements.txt
chromadb>=0.5.0
mem0ai>=0.1.0
requests>=2.31.0
sentence-transformers>=3.0.0